requirements.txt
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
Pillow==8.4.0
djangorestframework-simplejwt===5.0.0
django-environ===0.8.1
mysqlclient===2.1.0
dj-rest-auth===2.2.1
django-allauth===0.47.0
PyJWT===2.3.0
black
